Migration finished

+++ src/main/kotlin/io/tesfy/config/Allocation.kt
@@ -1,3 +1,3 @@
 package io.tesfy.config
 
-data class Allocation(val id: String, val rangeEnd: Double)
+data class Allocation(val id: String, val rangeEnd: Long)
