+++ src/main/kotlin/io/tesfy/config/Variation.kt
@@ -1,3 +1,3 @@
 package io.tesfy.config
 
-data class Variation (val id: String, val percentage: Double)
+data class Variation (val id: String, val percentage: Int)
